link-parser: Move get_line() to parser_utilities
