+++ link-parser/parser-utilities.c
@@ -37,6 +37,7 @@
 #include <string.h>
 
 #include "parser-utilities.h"
+#include "lg_readline.h"
 
 /**
  * Expand an initial '~' to home directory.
@@ -374,24 +375,91 @@ char **ms_windows_setup(int argc)
 
 	return argv;
 }
-
-char *get_utf8_line(char *input_string, int max_input_line, FILE *in)
-{
-	char *pline;
-
-	if (!running_under_cygwin)
-		pline = get_console_line();
-	else
-		pline = fgets(input_string, max_input_line, in);
-
-	return pline;
-}
-#else
-char *get_utf8_line(char *input_string, int max_input_line, FILE *in)
-{
-	return fgets(input_string, max_input_line, in);
-}
 #endif /* _WIN32 */
+
+/*
+ * @return -1: error; 0: EOF; 1: Success.
+ */
+static int fgets_with_check(char *inbuf, unsigned int inbuf_size, FILE *fh)
+{
+	const char *rc = fgets(inbuf, inbuf_size, fh);
+	if (rc == NULL)
+	{
+		/* EOF or error */
+		if (!ferror(fh)) return 0;
+		snprintf(inbuf, inbuf_size, "fgets(): %s", strerror(errno));
+		return -1;
+	}
+
+	size_t len = strlen(inbuf);
+	if ((len == inbuf_size -1) && inbuf[len -2] != '\n')
+	{
+		snprintf(inbuf, inbuf_size, "Input line too long (>%u).", inbuf_size-2);
+		return -1;
+	}
+
+	return 1;
+}
+
+static int get_terminal_line(const char *uprompt, char **buf,
+                             unsigned int bufsize, FILE *in, FILE *out, bool tty)
+{
+	int rc;
+
+#ifdef HAVE_EDITLINE
+	*buf = lg_readline(uprompt);
+	rc = (*buf != NULL);
+#else
+	fprintf(out, "%s", uprompt);
+	fflush(out);
+#ifdef _WIN32
+	if (!running_under_cygwin || tty)
+	{
+		rc = get_console_line(*buf, bufsize);
+	}
+	else
+#endif /* _WIN32 */
+	{
+		rc = fgets_with_check(*buf, bufsize, in);
+	}
+#endif /* HAVE_EDITLINE */
+
+	return rc;
+}
+
+/*
+ * Get an input line. Notify errors.
+ *
+ * @param in: Input file handle.
+ * @param out: Output file handle.
+ * @param tty \c false: from file; \c true: from terminal.
+ * @return \c true if successful, \c false on error.
+ */
+bool get_line(const char *uprompt, char **buf, unsigned int bufsize,
+                     FILE *in, FILE *out, bool tty)
+{
+	int rc;
+
+	if ((in != stdin) || !tty)
+	{
+		/* Get input from a file. */
+		rc = fgets_with_check(*buf, bufsize, in);
+	}
+	else
+	{
+		/* If we are here, the input is from a terminal. */
+		rc = get_terminal_line(uprompt, buf, bufsize, in, out, tty);
+	}
+
+	if (rc == 0) return false; /* EOF */
+	if (rc == -1)
+	{
+		prt_error("Fatal error: %s\n", *buf);
+		return false;
+	}
+
+	return true;
+}
 
 static unsigned int screen_width = INITIAL_SCREEN_WIDTH;
 /**
